add global word-break
